decision-agent/agent-backend/agent-executor/pyproject.toml
[project]
name = "agent-executor"
version = "0.1.0"
description = "Agent executor is an executor for agents"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiohttp==3.9.3",
    "apscheduler==3.11.0",
    "circuitbreaker==2.0.0",
    "deprecated>=1.3.1",
    "fastapi==0.119.1",
    "jsonschema==4.17.0",
    "limiter==0.5.0",
    "opentelemetry-instrumentation-aiohttp-client==0.61b0",
    "opentelemetry-sdk==1.40.0",
    "opentelemetry-exporter-otlp-proto-http==1.40.0",
    "pycryptodome==3.20.0",
    "pydantic==2.12.2",
    "python-multipart==0.0.22",
    "pyyaml==6.0.1",
    "redis==4.6.0",
    "requests==2.31.0",
    "sse-starlette>=3.0.3",
    "setuptools==76.0.0",
    "structlog==24.1.0",
    "setuptools==76.0.0",
    "uvicorn==0.38.0",
    #"mq-sdk@git+ssh://devops.aishu.cn:22/AISHUDevOps/One-Architecture/_git/proton-mq-python@1.7.4",
    #    "mq-sdk@git+https://github.com/kweaver-ai/proton-mq-sdk-py@release/1.9.0",
    #"telemetrysdk-python@git+ssh://devops.aishu.cn:22/AISHUDevOps/One-Architecture/_git/TelemetrySDK-Python@2.6.1",
    #    "telemetrysdk-python@git+https://github.com/kweaver-ai/TelemetrySDK-Python@2.4.2",
    #    "tlogging",
    "kweaver-dolphin==0.4.3",
    "python-dotenv>=1.0.0",
]


[dependency-groups]

dev = [
    "ruff>=0.14.1",
    "mypy>=1.0.0",
    "pyinstaller>=6.16.0",
    "pylint>=3.0.3",
]
 test = [
     "pytest>=7.0.0",
     "pytest-asyncio>=0.21.0",
     "pytest-cov>=4.0.0",
     "pylint==3.0.3",
     "pylint-junit==0.3.3",
 ]

lint = [
    "pre-commit>=4.5.1",
    "ruff>=0.15.0",
]
[tool.uv]
index-url = "https://pypi.tuna.tsinghua.edu.cn/simple"


#[[tool.uv.index]]
#name = "test-pypi"
#url = "https://test.pypi.org/simple/"

##  打开下方注释则从 TestPyPI 安装（测试版本）
#index-strategy = "unsafe-best-match"
#[tool.uv.sources]
# kweaver-dolphin 源切换配置
#kweaver-dolphin = { index = "test-pypi" }

#mq-sdk = { git = "https://github.com/kweaver-ai/proton-mq-sdk-py", rev = "release/1.9.0" }
#tlogging = { git = "https://github.com/kweaver-ai/TelemetrySDK-Python", rev = "2.4.2" }

[tool.pytest.ini_options]
norecursedirs = ["test/unit/common/tool_v2", "test/unit/boot"]
asyncio_mode = "auto"


[tool.coverage.run]
omit = [
    "main.py",
    "*/test/*",           # 不统计测试文件本身
    "*/test_*",
    "test/unit/common/tool_v2",
    ".venv/*",              # 跳过虚拟环境
    "*/__pycache__/*",
]

[tool.ruff]
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W"]
ignore = [
    "E501",  # line too long
    "E722",  # bare except
    "W605",  # invalid escape sequence (regex patterns)
]

[tool.ruff.lint.per-file-ignores]
"test/*" = [
    "F401",  # unused import
    "F811",  # redefinition of unused
    "F821",  # undefined name
    "F841",  # unused variable
    "E721",  # type comparison
    "E731",  # lambda assignment
    "E741",  # ambiguous variable name
    "E402",  # import not at top
]
